remove AdaptiveBeamLengthConstraint from .pro
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/beamadapter/BeamAdapter@23 f2fb72c6-44b1-464a-b922-67ecee8ddb02

+++ BeamInterpolation.h
@@ -260,7 +260,7 @@ public:
 
         void setTransformBetweenDofAndNode(int beam, const Transform &DOF_H_Node, unsigned int zeroORone )
         {
-            if(beam > (DOF0_Transform_node0.size()-1) || beam > (DOF1_Transform_node1.size()-1) )
+            if(beam > (int) (DOF0_Transform_node0.size()-1) || beam > (int) (DOF1_Transform_node1.size()-1) )
             {
                 serr<<"WARNING setTransformBetweenDofAndNode on non existing beam"<<sendl;
                 return;
